fix(chat): drop error-text regex from title sanitiser, share length cap
  The refusal/error regex in generate_title rejected legitimate titles
  ("Warning signs of burnout", "Error culture in hospitals") and missed
  most real failures: non-English refusals, curly apostrophes, and
  multi-line error text whose "Error:" header the cleaner already
  stripped. Remove it and keep only the length cap.

  - Move the cap into dembrane/utils.py as sanitize_generated_title so
    chat titles and conversation titles share one guard
  - Run the sanitiser inside the try block so nothing in generate_title
    can fail the chat turn
  - Log title failures with logger.exception so a persistent model outage
    still reaches Sentry instead of one warning per turn
  - Tests: patch DISABLE_CHAT_TITLE_GENERATION so they cannot pass
    vacuously, reuse tests/llm_fakes.py, and let the fakes carry None
    content
